When PR created no need to have data cleanup

+++ tests/e2e/features/test_e2e_build_application_questions_from_create_grants.py
@@ -69,5 +69,6 @@ def test_e2e_grant_creation_to_application_completion_flow(page: Page, domains: 
             .then_validate_download_success()
         )
     finally:
-        HttpClient().delete(output, "grants")
-        HttpClient().delete(output, "templates")
+        if domains.environment != "e2e":
+            HttpClient().delete(output, "grants")
+            HttpClient().delete(output, "templates")
